clean up clippy and build errors
