+++ src/model/ki2600.rs
@@ -472,7 +472,10 @@ mod unit {
 
         let mut instrument: Instrument = Instrument::new(
             protocol::Protocol::Raw(Raw::new(interface)),
-            Authentication::NoAuth,
+            Authentication::Credential {
+                username: String::new(),
+                password: "secret_token".to_string(),
+            },
         );
 
         assert_matches!(instrument.check_login(), Ok(instrument::State::Needed));
@@ -615,7 +618,10 @@ mod unit {
             .returning(|buf: &[u8]| Ok(buf.len()));
         let mut instrument: Instrument = Instrument::new(
             protocol::Protocol::Raw(Raw::new(interface)),
-            Authentication::NoAuth,
+            Authentication::Credential {
+                username: String::new(),
+                password: "secret_token".to_string(),
+            },
         );
 
         assert_matches!(instrument.check_login(), Ok(instrument::State::Needed));
