+++ src/model/ki3700.rs
@@ -473,7 +473,10 @@ mod unit {
 
         let mut instrument: Instrument = Instrument::new(
             protocol::Protocol::Raw(Raw::new(interface)),
-            Authentication::NoAuth,
+            Authentication::Credential {
+                username: String::new(),
+                password: "secret_token".to_string(),
+            },
         );
 
         assert_matches!(instrument.check_login(), Ok(instrument::State::Needed));
@@ -617,7 +620,10 @@ mod unit {
             .returning(|buf: &[u8]| Ok(buf.len()));
         let mut instrument: Instrument = Instrument::new(
             protocol::Protocol::Raw(Raw::new(interface)),
-            Authentication::NoAuth,
+            Authentication::Credential {
+                username: String::new(),
+                password: "secret_token".to_string(),
+            },
         );
 
         assert_matches!(instrument.check_login(), Ok(instrument::State::Needed));
